+++ spec/idempotent_spec.rb
@@ -162,7 +162,7 @@ describe Grape::Idempotency do
                 header "idempotency-key", idempotency_key
                 post 'payments?locale=en', { amount: 800_00 }.to_json
                 expect(last_response.status).to eq(409)
-                expect(last_response.body).to eq({ "error" => "You are using the same idempotent key for two different requests" }.to_json)
+                expect(last_response.body).to eq("{\"error\"=>\"You are using the same idempotent key for two different requests\"}")
               end
             end
 
@@ -192,7 +192,7 @@ describe Grape::Idempotency do
                 header "idempotency-key", idempotency_key
                 post 'refunds?locale=es', { amount: 100_00 }.to_json
                 expect(last_response.status).to eq(409)
-                expect(last_response.body).to eq({ "error" => "You are using the same idempotent key for two different requests" }.to_json)
+                expect(last_response.body).to eq("{\"error\"=>\"You are using the same idempotent key for two different requests\"}")
               end
             end
           end
@@ -240,12 +240,42 @@ describe Grape::Idempotency do
                 header "idempotency-key", idempotency_key
                 post 'payments?locale=es', { amount: 100_00 }.to_json
                 expect(last_response.status).to eq(500)
-                expect(last_response.body).to eq("{\"error\":\"Internal Server Error\",\"message\":\"Unexpected error\"}")
+                expect(last_response.body).to eq("{:error=>\"Internal Server Error\", :message=>\"Unexpected error\"}")
   
                 header "idempotency-key", idempotency_key
                 post 'payments?locale=es', { amount: 100_00 }.to_json
                 expect(last_response.status).to eq(500)
-                expect(last_response.body).to eq("{\"error\":\"Internal Server Error\",\"message\":\"Unexpected error\"}")
+                expect(last_response.body).to eq("{\"error\"=>\"Internal Server Error\", \"message\"=>\"Unexpected error\"}")
+              end
+            end
+
+            context 'and a managed exception using grape rescue_from appears executing the code' do
+              it 'stores the exception response and returns the same response in the second call' do
+                allow(SecureRandom).to receive(:random_number).and_return(1, 2)
+
+                app.rescue_from StandardError do |e|
+                  status = 404
+                  error = { message: "Not found error" }
+                  error!(error, status)
+                end
+
+                app.post('/payments') do
+                  idempotent do
+                    raise "Not found error error" if SecureRandom.random_number == 1
+                    status 200
+                    { amount_to: 100_00 }.to_json
+                  end
+                end
+  
+                header "idempotency-key", idempotency_key
+                post 'payments?locale=es', { amount: 100_00 }.to_json
+                expect(last_response.status).to eq(404)
+                expect(last_response.body).to eq("{\"message\":\"Not found error\"}")
+  
+                header "idempotency-key", idempotency_key
+                post 'payments?locale=es', { amount: 100_00 }.to_json
+                expect(last_response.status).to eq(404)
+                expect(last_response.body).to eq("{\"message\"=>\"Not found error\"}")
               end
             end
 
@@ -470,7 +500,7 @@ describe Grape::Idempotency do
         header "idempotency-key", idempotency_key
         post 'payments?locale=en', { amount: 800_00 }.to_json
         expect(last_response.status).to eq(409)
-        expect(last_response.body).to eq(expected_error_response.to_json)
+        expect(last_response.body).to eq("{:error=>\"An error wadus with conflict\", :status=>409, :message=>\"You are using the same idempotency key for two different requests\"}")
       end
     end
   end