+++ CHANGELOG.md
@@ -4,12 +4,18 @@ All changes to `grape-idempotency` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2023-01-07
+
+### Fix
+
+- Second calls were returning `null` when the first response was generated inside a `rescue_from`.
+- Conflict response had invalid format.
+
 ## [0.1.2] - 2023-01-06
 
 ### Fix
 
 - Return correct original response when the endpoint returns a hash in the body
-
 
 ## [0.1.1] - 2023-01-06